fix(createThemeConfig): Support RSC (#82)
* feat: create-theme-config ssr 지원

* docs: themeing 사용법 변경

* chore: create changeset

* chore: update version

* docs: next.js 설정 변경
