+++ .changeset/brave-forks-appear.md
@@ -1,0 +1,6 @@
+---
+'@vapor-ui/core': patch
+'website': patch
+---
+
+fix(createThemeConfig): support RSC
