+++ packages/core/src/components/create-theme-config/index.ts
@@ -1,0 +1,1 @@
+export * from './create-theme-config';
